Gave Tests and Demos their own Class; Changed shuffletest so you can imput the amount of numbers to shuffle; changed AnyDemo with a better example;
